IT'S SPOOKY MONTH

+++ commands/mod/reboot.js
@@ -1,4 +1,7 @@
-module.exports = (client, message) => {
+module.exports = (client, message, config) => {
+    let mod = message.author
+    if (mod.id == config.owner[1] || mod.id == config.wife[1]) return
+
     message.channel.send("Reiniciando.. por favor espere 10s antes de usar un comando de música")
     .then(() => client.destroy())
     .then(() => client.login(process.env.TOKEN))
